+++ engine.go
@@ -68,20 +68,12 @@ type Engine struct {
 	noScheduler bool
 }
 
-// handlerEntry pairs a stable ID with a message handler function.
-type handlerEntry struct {
-	id      HandlerID
-	handler MessageHandler
-}
-
 // runHandle tracks an in-flight workflow execution.
 type runHandle struct {
 	cancel   context.CancelFunc
 	done     chan *Result
 	finished chan struct{} // closed when all persistence writes are done
 }
-
-const defaultCacheSize = 1000
 
 // EngineOption configures the engine.
 type EngineOption func(*Engine) error
@@ -461,81 +453,6 @@ func (e *Engine) Resume(ctx context.Context, wf *Workflow, runID RunID, data Par
 	return result, nil
 }
 
-// OnMessage registers a handler for IPC messages of the given type.
-// Handlers fire when a child process sends a message via Send(),
-// or when Send() is called in the parent process (local routing).
-// Returns a HandlerID that can be passed to OffMessage to deregister.
-func (e *Engine) OnMessage(msgType string, handler MessageHandler) HandlerID {
-	if handler == nil {
-		return 0
-	}
-	id := HandlerID(e.messageHandlerCounter.Add(1))
-	e.messageHandlersMu.Lock()
-	defer e.messageHandlersMu.Unlock()
-	e.messageHandlers[msgType] = append(e.messageHandlers[msgType], handlerEntry{id: id, handler: handler})
-	e.messageHandlerIndex[id] = msgType
-	return id
-}
-
-// OffMessage removes a previously registered handler by its HandlerID.
-// If the ID is not found, it is a no-op.
-func (e *Engine) OffMessage(id HandlerID) {
-	if id == 0 {
-		return
-	}
-	e.messageHandlersMu.Lock()
-	defer e.messageHandlersMu.Unlock()
-
-	msgType, ok := e.messageHandlerIndex[id]
-	if !ok {
-		return
-	}
-	delete(e.messageHandlerIndex, id)
-
-	entries := e.messageHandlers[msgType]
-	for i, entry := range entries {
-		if entry.id == id {
-			newEntries := make([]handlerEntry, 0, len(entries)-1)
-			newEntries = append(newEntries, entries[:i]...)
-			newEntries = append(newEntries, entries[i+1:]...)
-			e.messageHandlers[msgType] = newEntries
-			return
-		}
-	}
-}
-
-// dispatchMessage routes an IPC message to all registered handlers.
-func (e *Engine) dispatchMessage(msgType string, payload map[string]any) {
-	e.messageHandlersMu.RLock()
-	entries := e.messageHandlers[msgType]
-	// Also fire wildcard handlers (registered with "*")
-	wildcardEntries := e.messageHandlers["*"]
-	e.messageHandlersMu.RUnlock()
-
-	for _, entry := range entries {
-		fn := entry.handler
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					e.logger.Error("message handler panicked for %q: %v", msgType, r)
-				}
-			}()
-			fn(msgType, payload)
-		}()
-	}
-	for _, entry := range wildcardEntries {
-		fn := entry.handler
-		func() {
-			defer func() {
-				if r := recover(); r != nil {
-					e.logger.Error("wildcard message handler panicked for %q: %v", msgType, r)
-				}
-			}()
-			fn(msgType, payload)
-		}()
-	}
-}
-
 // Close releases engine resources with ordered shutdown.
 // Safe to call multiple times — the second and subsequent calls are no-ops.
 func (e *Engine) Close() error {
@@ -574,49 +491,6 @@ func (e *Engine) Close() error {
 		closeErr = e.persistence.Close()
 	})
 	return closeErr
-}
-
-// cacheWorkflow stores a workflow in the engine's cache for timer-based recovery.
-// Uses O(1) LRU eviction via a doubly-linked list + map node index.
-// Front of the list = most recently used; back = least recently used.
-func (e *Engine) cacheWorkflow(wf *Workflow) {
-	e.workflowCacheMu.Lock()
-	defer e.workflowCacheMu.Unlock()
-
-	if node, exists := e.workflowCacheNode[wf.ID]; exists {
-		// Cache hit: promote to front (O(1)).
-		e.workflowCacheList.MoveToFront(node)
-	} else {
-		// New entry: evict LRU if at capacity (skip pinned entries).
-		if e.cacheSize > 0 && len(e.workflowCache) >= e.cacheSize {
-			evicted := false
-			for el := e.workflowCacheList.Back(); el != nil; el = el.Prev() {
-				id := el.Value.(string)
-				if e.pinnedWorkflows[id] == 0 {
-					e.workflowCacheList.Remove(el)
-					delete(e.workflowCacheNode, id)
-					delete(e.workflowCache, id)
-					evicted = true
-					break
-				}
-			}
-			if !evicted {
-				// All entries are pinned — allow cache to grow beyond limit.
-				// This is safer than losing a sleeping workflow.
-				e.logger.Warn("workflow cache exceeds limit: all %d entries are pinned by sleeping runs", len(e.workflowCache))
-			}
-		}
-		node := e.workflowCacheList.PushFront(wf.ID)
-		e.workflowCacheNode[wf.ID] = node
-	}
-	e.workflowCache[wf.ID] = wf
-}
-
-// lookupCachedWorkflow retrieves a workflow from the cache.
-func (e *Engine) lookupCachedWorkflow(id string) *Workflow {
-	e.workflowCacheMu.RLock()
-	defer e.workflowCacheMu.RUnlock()
-	return e.workflowCache[id]
 }
 
 // wakeWorkflow is called by the timer scheduler when a sleeping workflow's timer fires.
